refactor: 모듈화 완료
