+++ weekly-mission(week2)/signup.html
@@ -58,6 +58,6 @@
       </footer>
     </div>
   </div>
-  <script type="module" src="./src/login.js"></script>
+  <script type="module" src="./src/features/userManagement.js"></script>
 </body>
 </html>
