+++ weekly-mission(week2)/src/constants/accounts.js
@@ -1,0 +1,6 @@
+const MASTER_ACCOUNT = Object.freeze({
+  EMAIL: 'test@codeit.com',
+  PASSWORD: 'codeit101'
+});
+
+export default MASTER_ACCOUNT;
