+++ weekly-mission(week2)/src/constants/validate.js
@@ -1,0 +1,6 @@
+import REGEX  from "./regex.js";
+
+export const isValid = {
+  email: (email) => REGEX.EMAIL.test(email),
+  password: (password) => REGEX.PASSWORD.test(password)
+};
